+++ src/Middleware/PsrJsonApiMiddleware.php
@@ -112,15 +112,18 @@ class PsrJsonApiMiddleware implements MiddlewareInterface
     {
         if ($data = file_get_contents($this->streamPath)) {
             $body = json_decode($data);
-            $result = self::$validator->schemaValidation($body, self::$schema);
             if (json_last_error() !== JSON_ERROR_NONE) {
                 throw new BadRequest(json_last_error_msg());
             }
-            if (!$result->isValid()) {
-                $error = $result->getFirstError();
-                $msg = "Error: " . $error->keyword() . ' args: [' . print_r($error->keywordArgs()) . ']';
-                throw new BadRequest($msg);
-            }
+//          todo: Disabled cause weird behavior see
+//                https://discuss.jsonapi.org/t/json-schema-required-id-in-case-of-creating-resourceobject/1724
+//
+//            $result = self::$validator->schemaValidation($body, self::$schema);
+//            if (!$result->isValid()) {
+//                $error = $result->getFirstError();
+//                $msg = "Error: " . $error->keyword() . ' args: [' . print_r($error->keywordArgs(), true) . ']';
+//                throw new BadRequest($msg);
+//            }
             return $body;
         }
         return null;